[anhalyze_geo] Moving deprecated function to deprectated file, making cleaner transition to new version.
